requirements.txt
pubnub==4.1.2
pyserial==3.4
python-dotenv==0.10.1
